+++ Spaxel/src/code/entity/Particle.java
@@ -5,12 +5,12 @@ import code.graphics.RenderBuffer;
 import code.graphics.Sprite;
 
 public class Particle extends Entity{
-	private double deltaRot;
-	private double dir;
-	private double speed;
+	private float deltaRot;
+	private float dir;
+	private float speed;
 	private Sprite sprite;
 	
-	public Particle(double x, double y, double rot, double deltaRot, double dir, double speed, int life, Sprite sprite){
+	public Particle(float x, float y, float rot, float deltaRot, float dir, float speed, int life, Sprite sprite){
 		super(x, y, rot);
 		this.deltaRot = deltaRot;
 		this.dir = dir;
@@ -21,8 +21,8 @@ public class Particle extends Entity{
 	
 	public void update(){
 		rot += deltaRot;
-		double dx = Math.sin(dir) * speed;
-		double dy = Math.cos(dir) * speed;
+		float dx = (float)Math.sin(dir) * speed;
+		float dy = (float)Math.cos(dir) * speed;
 		x -= dx* Engine.getEngine().getUpdateTime();
 		y -= dy*Engine.getEngine().getUpdateTime();
 	}
