+++ Spaxel/src/code/factories/BasicLaserFactory.java
@@ -7,11 +7,11 @@ public class BasicLaserFactory extends ProjectileFactory{
 	
 
 	
-	public BasicLaserFactory(Sprite sprite,Sprite trail, int damage, int life, double speed){
+	public BasicLaserFactory(Sprite sprite,Sprite trail, int damage, int life, float speed){
 		super(sprite,trail, damage, life, speed);
 	}
 	
-	public BasicLaser make(double x, double y, double rot){
+	public BasicLaser make(float x, float y, float rot){
 		return new BasicLaser(x,y, rot, sprite, trail,damage, life, speed);
 	}
