+++ Spaxel/src/code/factories/HackingMissileFactory.java
@@ -8,11 +8,11 @@ import code.projectiles.Projectile;
  * Created by theo on 13-5-2016.
  */
 public class HackingMissileFactory extends ProjectileFactory {
-    public HackingMissileFactory(Sprite sprite,Sprite trail,  int damage, int life, double speed) {
+    public HackingMissileFactory(Sprite sprite,Sprite trail,  int damage, int life, float speed) {
         super(sprite, trail, damage, life, speed);
     }
 
-    public Projectile make(double x, double y, double rot){
+    public Projectile make(float x, float y, float rot){
         return new HackingMissile(x, y, rot, sprite, trail, damage, life, speed);
     }
 }
