+++ Spaxel/src/code/projectiles/BasicLaser.java
@@ -4,7 +4,7 @@ import code.graphics.Sprite;
 
 public class BasicLaser extends Projectile {
 
-	public BasicLaser(double x, double y, double rot, Sprite sprite, Sprite trail, int damage, int life, double speed) {
+	public BasicLaser(float x, float y, float rot, Sprite sprite, Sprite trail, int damage, int life, float speed) {
 		super(x, y, rot, sprite, trail, damage, life, speed);
 	}
 }
