+++ Spaxel/src/code/projectiles/BasicMissile.java
@@ -4,7 +4,7 @@ import code.graphics.Sprite;
 
 public class BasicMissile extends Projectile{
 
-	public BasicMissile(double x, double y, double rot, Sprite sprite, Sprite trail, int damage, int life, double speed) {
+	public BasicMissile(float x, float y, float rot, Sprite sprite, Sprite trail, int damage, int life, float speed) {
 		super(x, y, rot, sprite,trail, damage, life, speed);
 	}
