+++ Spaxel/src/code/projectiles/DisruptLaser.java
@@ -9,7 +9,7 @@ import code.inventory.ShootEffect;
  * Created by theo on 12-5-2016.
  */
 public class DisruptLaser extends Projectile {
-    public DisruptLaser(double x, double y, double rot, Sprite sprite,Sprite trail,  int damage, int life, double speed) {
+    public DisruptLaser(float x, float y, float rot, Sprite sprite,Sprite trail,  int damage, int life, float speed) {
         super(x, y, rot, sprite,trail, damage, life, speed);
     }
